KaryTree WIP (massive commit). Got some design decisions to make. See full commit log for details.
There are some issues with timing as far as asserting that a tree will be
unary, for example. Imagine two clients connecting to the broadcaster at
once, when K=1. We'd like to have exactly one disconnect, and have it
connect to the other node.

Then the clients could randomly decide to disconnect, but then you run
into the problem of both disconnecting at the same time.

Alternatively, you might have the broadcaster instruct one to disconnect,
but what if there's a network failure between the two nodes so that they
have to connect to the broadcaster?

Right now, at the very least, nodes don't even know to disconnect and
leave the broadcaster alone until a failure arises.